refactor: change how move secrets happen, so this can be reused withou some validations on move folders
