+++ backend/src/services/secret-v2-bridge/secret-v2-bridge-service.ts
@@ -26,10 +26,6 @@ import { ProjectEvents } from "@app/ee/services/project-events/project-events-ty
 import { TSecretApprovalPolicyServiceFactory } from "@app/ee/services/secret-approval-policy/secret-approval-policy-service";
 import { TSecretApprovalRequestDALFactory } from "@app/ee/services/secret-approval-request/secret-approval-request-dal";
 import { TSecretApprovalRequestSecretDALFactory } from "@app/ee/services/secret-approval-request/secret-approval-request-secret-dal";
-import {
-  InternalMetadataType,
-  TInternalMetadata
-} from "@app/ee/services/secret-approval-request/secret-approval-request-types";
 import { scanSecretPolicyViolations } from "@app/ee/services/secret-scanning-v2/secret-scanning-v2-fns";
 import { TSecretSnapshotServiceFactory } from "@app/ee/services/secret-snapshot/secret-snapshot-service";
 import { KeyStorePrefixes, KeyStoreTtls, TKeyStoreFactory } from "@app/keystore/keystore";
@@ -40,7 +36,6 @@ import { BadRequestError, ForbiddenRequestError, NotFoundError } from "@app/lib/
 import { diff, groupBy } from "@app/lib/fn";
 import { setKnexStringValue } from "@app/lib/knex";
 import { logger } from "@app/lib/logger";
-import { alphaNumericNanoId } from "@app/lib/nanoid";
 import { requestMemoKeys } from "@app/lib/request-context/memo-keys";
 import { requestMemoize } from "@app/lib/request-context/request-memoizer";
 import {
@@ -85,20 +80,20 @@ import {
   fnSecretBulkDelete,
   fnSecretBulkInsert,
   fnSecretBulkUpdate,
-  fnUpdateMovedSecretReferences,
+  fnSecretMoveInTransaction,
   fnUpdateSecretLinkedReferences,
   generatePaths,
   recursivelyGetSecretPaths,
   reshapeBridgeSecret
 } from "./secret-v2-bridge-fns";
 import {
-  SecretOperations,
   SecretUpdateMode,
   TBackFillSecretReferencesDTO,
   TCreateManySecretDTO,
   TCreateSecretDTO,
   TDeleteManySecretDTO,
   TDeleteSecretDTO,
+  TDispatchSecretMoveSideEffectsDTO,
   TGetAccessibleSecretsDTO,
   TGetASecretDTO,
   TGetSecretReferencesTreeDTO,
@@ -2969,501 +2964,21 @@ export const secretV2BridgeServiceFactory = ({
     return { message: "Successfully backfilled secret references" };
   };
 
-  const moveSecrets = async ({
-    sourceEnvironment,
-    sourceSecretPath,
-    destinationEnvironment,
-    destinationSecretPath,
-    secretIds,
+  const dispatchSecretMoveSideEffects = async ({
     projectId,
-    shouldOverwrite,
+    orgId,
     actor,
     actorId,
-    actorAuthMethod,
-    actorOrgId,
-    skipPermissionCheck
-  }: TMoveSecretsDTO) => {
-    const { permission } = await permissionService.getProjectPermission({
-      actor,
-      actorId,
-      projectId,
-      actorAuthMethod,
-      actorOrgId,
-      actionProjectType: ActionProjectType.SecretManager
-    });
-
-    const sourceFolder = await folderDAL.findBySecretPath(projectId, sourceEnvironment, sourceSecretPath);
-    if (!sourceFolder) {
-      throw new NotFoundError({
-        message: `Source folder with path '${sourceSecretPath}' in environment with slug '${sourceEnvironment}' not found`
-      });
-    }
-
-    const destinationFolder = await folderDAL.findBySecretPath(
-      projectId,
-      destinationEnvironment,
-      destinationSecretPath
-    );
-
-    if (!destinationFolder) {
-      throw new NotFoundError({
-        message: `Destination folder with path '${destinationSecretPath}' in environment with slug '${destinationEnvironment}' not found`
-      });
-    }
-
-    const sourceSecrets = await secretDAL.find({
-      type: SecretType.Shared,
-      folderId: sourceFolder.id,
-      $in: {
-        [`${TableName.SecretV2}.id` as "id"]: secretIds
-      }
-    });
-
-    if (sourceSecrets.length !== secretIds.length) {
-      throw new NotFoundError({
-        message: `One or more secrets not found in source folder with path '${sourceSecretPath}' and environment slug '${sourceEnvironment}'`
-      });
-    }
-    const sourceActions = [
-      ProjectPermissionSecretActions.Delete,
-      ProjectPermissionSecretActions.ReadValue,
-      ProjectPermissionSecretActions.DescribeSecret
-    ] as const;
-    const destinationActions = [ProjectPermissionSecretActions.Create, ProjectPermissionSecretActions.Edit] as const;
-
-    sourceSecrets.forEach((secret) => {
-      if (secret.isRotatedSecret) {
-        throw new BadRequestError({ message: `Cannot move rotated secret: ${secret.key}` });
-      }
-      if (secret.isHoneyTokenSecret) {
-        throw new BadRequestError({ message: `Cannot move honey token secret: ${secret.key}` });
-      }
-
-      // the caller (e.g. folder move) has already pre-authorized the move at the path level
-      if (skipPermissionCheck) return;
-
-      for (const sourceAction of sourceActions) {
-        if (
-          sourceAction === ProjectPermissionSecretActions.DescribeSecret ||
-          sourceAction === ProjectPermissionSecretActions.ReadValue
-        ) {
-          throwIfMissingSecretReadValueOrDescribePermission(permission, sourceAction, {
-            environment: sourceEnvironment,
-            secretPath: sourceSecretPath,
-            secretName: secret.key,
-            secretTags: secret.tags.map((el) => el.slug)
-          });
-        } else {
-          ForbiddenError.from(permission).throwUnlessCan(
-            sourceAction,
-            subject(ProjectPermissionSub.Secrets, {
-              environment: sourceEnvironment,
-              secretPath: sourceSecretPath,
-              secretName: secret.key,
-              secretTags: secret.tags.map((el) => el.slug)
-            })
-          );
-        }
-      }
-    });
-
-    if (sourceSecrets.length !== secretIds.length) {
-      throw new BadRequestError({
-        message: "Invalid secrets"
-      });
-    }
-
-    const {
-      encryptor: secretManagerEncryptor,
-      decryptor: secretManagerDecryptor,
-      generateSecretBlindIndex
-    } = await kmsService.createCipherPairWithDataKey({
-      type: KmsDataKey.SecretManager,
-      projectId
-    });
-    const decryptedSourceSecrets = sourceSecrets.map((secret) => ({
-      ...secret,
-      value: secret.encryptedValue
-        ? secretManagerDecryptor({ cipherTextBlob: secret.encryptedValue }).toString()
-        : undefined
-    }));
-
-    let isSourceUpdated = false;
-    let isDestinationUpdated = false;
-
-    // Moving secrets is a two-step process.
-    await secretDAL.transaction(async (tx) => {
-      // First step is to create/update the secret in the destination:
-      const destinationSecretsFromDB = await secretDAL.find(
-        {
-          folderId: destinationFolder.id
-        },
-        { tx }
-      );
-
-      const decryptedDestinationSecrets = destinationSecretsFromDB.map((secret) => {
-        return {
-          ...secret,
-          value: secret.encryptedValue
-            ? secretManagerDecryptor({ cipherTextBlob: secret.encryptedValue }).toString()
-            : undefined
-        };
-      });
-
-      const destinationSecretsGroupedByKey = groupBy(decryptedDestinationSecrets, (i) => i.key);
-
-      const sourceKeys = decryptedSourceSecrets.map((s) => s.key);
-
-      const conflictingRotationSecretKeys = sourceKeys.filter(
-        (key) => destinationSecretsGroupedByKey[key]?.[0]?.isRotatedSecret
-      );
-      if (conflictingRotationSecretKeys.length > 0) {
-        throw new BadRequestError({
-          message: `Cannot move secrets to '${destinationFolder.path}' because the following keys are managed by a secret rotation at the destination: ${conflictingRotationSecretKeys.join(", ")}`
-        });
-      }
-
-      const conflictingHoneyTokenSecretKeys = sourceKeys.filter(
-        (key) => destinationSecretsGroupedByKey[key]?.[0]?.isHoneyTokenSecret
-      );
-      if (conflictingHoneyTokenSecretKeys.length > 0) {
-        throw new BadRequestError({
-          message: `Cannot move secrets to '${destinationFolder.path}' because the following keys are managed by a honey token at the destination: ${conflictingHoneyTokenSecretKeys.join(", ")}`
-        });
-      }
-
-      const locallyCreatedSecrets = decryptedSourceSecrets
-        .filter(({ key }) => !destinationSecretsGroupedByKey[key]?.[0])
-        .map((el) => ({ ...el, operation: SecretOperations.Create }));
-
-      const locallyUpdatedSecrets = decryptedSourceSecrets
-        .filter(
-          ({ key, value }) =>
-            destinationSecretsGroupedByKey[key]?.[0] && destinationSecretsGroupedByKey[key]?.[0]?.value !== value
-        )
-        .map((el) => ({ ...el, operation: SecretOperations.Update }));
-
-      if (locallyUpdatedSecrets.length > 0 && !shouldOverwrite) {
-        const existingKeys = locallyUpdatedSecrets.map((s) => s.key);
-
-        throw new BadRequestError({
-          message: `Failed to move secrets. The following secrets already exist in the destination: ${existingKeys.join(
-            ","
-          )}`
-        });
-      }
-
-      const isEmpty = locallyCreatedSecrets.length + locallyUpdatedSecrets.length === 0;
-
-      if (isEmpty) {
-        throw new BadRequestError({
-          message: "Selected secrets already exist in the destination."
-        });
-      }
-
-      // permission check whether can create or edit the ones in the destination folder
-      // (skipped when the caller has already pre-authorized the move at the path level)
-      if (!skipPermissionCheck) {
-        locallyCreatedSecrets.forEach((secret) => {
-          for (const destinationAction of destinationActions) {
-            ForbiddenError.from(permission).throwUnlessCan(
-              destinationAction,
-              subject(ProjectPermissionSub.Secrets, {
-                environment: destinationEnvironment,
-                secretPath: destinationFolder.path,
-                secretName: secret.key,
-                secretTags: secret.tags.map((el) => el.slug)
-              })
-            );
-          }
-        });
-      }
-
-      const destinationFolderPolicy = await secretApprovalPolicyService.getSecretApprovalPolicy(
-        projectId,
-        destinationFolder.environment.slug,
-        destinationFolder.path
-      );
-
-      let destinationSecretIdByKey: Record<string, string | undefined> = {};
-
-      if (destinationFolderPolicy && actor === ActorType.USER) {
-        // if secret approval policy exists for destination, we create the secret approval request
-        const localSecretsIds = decryptedDestinationSecrets.map(({ id }) => id);
-        const latestSecretVersions = await secretVersionDAL.findLatestVersionMany(
-          destinationFolder.id,
-          localSecretsIds,
-          tx
-        );
-
-        const approvalRequestDoc = await secretApprovalRequestDAL.create(
-          {
-            folderId: destinationFolder.id,
-            slug: alphaNumericNanoId(),
-            policyId: destinationFolderPolicy.id,
-            status: "open",
-            hasMerged: false,
-            committerUserId: actorId
-          },
-          tx
-        );
-
-        const secretsForApproval = locallyCreatedSecrets.concat(locallyUpdatedSecrets);
-        const commits = secretsForApproval.map((doc) => {
-          const { operation } = doc;
-          const localSecret = destinationSecretsGroupedByKey[doc.key]?.[0];
-
-          return {
-            ...(operation === SecretOperations.Create
-              ? {
-                  internalMetadata: {
-                    type: InternalMetadataType.MoveSecret,
-                    payload: {
-                      source: {
-                        secretPath: sourceSecretPath,
-                        environment: sourceEnvironment
-                      }
-                    }
-                  } as TInternalMetadata
-                }
-              : {}),
-            op: operation,
-            requestId: approvalRequestDoc.id,
-            metadata: doc.metadata,
-            key: doc.key,
-            encryptedValue: doc.encryptedValue,
-            encryptedComment: doc.encryptedComment,
-            skipMultilineEncoding: doc.skipMultilineEncoding,
-            // except create operation other two needs the secret id and version id
-            ...(operation !== SecretOperations.Create
-              ? { secretId: localSecret.id, secretVersion: latestSecretVersions[localSecret.id].id }
-              : {})
-          };
-        });
-        const approvalCommits = await secretApprovalRequestSecretDAL.insertV2Bridge(commits, tx);
-
-        const approvalCommitsGroupedByKey = groupBy(approvalCommits, (i) => i.key);
-        const approvalSecretTags = secretsForApproval.flatMap((doc) =>
-          doc.tags.map((tag) => ({
-            secretId: approvalCommitsGroupedByKey[doc.key][0].id,
-            tagId: tag.id
-          }))
-        );
-        if (approvalSecretTags.length) {
-          await secretApprovalRequestSecretDAL.insertApprovalSecretV2Tags(approvalSecretTags, tx);
-        }
-      } else {
-        // apply changes directly
-        let createdSecrets: { id: string; key: string }[] = [];
-
-        if (locallyCreatedSecrets.length) {
-          const inputSecretsForCreate = await Promise.all(
-            locallyCreatedSecrets.map(async (doc) => ({
-              type: doc.type,
-              metadata: doc.metadata,
-              key: doc.key,
-              encryptedValue: doc.encryptedValue,
-              encryptedComment: doc.encryptedComment,
-              skipMultilineEncoding: doc.skipMultilineEncoding,
-              reminderNote: doc.reminderNote,
-              reminderRepeatDays: doc.reminderRepeatDays,
-              secretMetadata: doc.secretMetadata?.map(({ key, value, encryptedValue }) => ({
-                key,
-                value: value || undefined,
-                encryptedValue: encryptedValue || undefined
-              })) as { key: string; value?: string; encryptedValue?: Buffer }[] | undefined,
-              references: doc.value ? getAllSecretReferences(doc.value).nestedReferences : [],
-              tagIds: doc.tags.map((tag) => tag.id),
-              secretValueBlindIndex: doc.value ? await generateSecretBlindIndex(Buffer.from(doc.value)) : undefined
-            }))
-          );
-
-          createdSecrets = await fnSecretBulkInsert({
-            folderId: destinationFolder.id,
-            orgId: actorOrgId,
-            secretVersionDAL,
-            secretDAL,
-            tx,
-            secretTagDAL,
-            resourceMetadataDAL,
-            folderCommitService,
-            secretVersionTagDAL,
-            actor: {
-              type: actor,
-              actorId
-            },
-            inputSecrets: inputSecretsForCreate
-          });
-        }
-        if (locallyUpdatedSecrets.length) {
-          const inputSecretsForUpdate = await Promise.all(
-            locallyUpdatedSecrets.map(async (doc) => ({
-              filter: {
-                folderId: destinationFolder.id,
-                id: destinationSecretsGroupedByKey[doc.key][0].id
-              },
-              data: {
-                metadata: doc.metadata,
-                key: doc.key,
-                encryptedComment: doc.encryptedComment,
-                skipMultilineEncoding: doc.skipMultilineEncoding,
-                secretMetadata: doc.secretMetadata?.map(({ key, value, encryptedValue }) => ({
-                  key,
-                  value,
-                  encryptedValue
-                })) as { key: string; value?: string; encryptedValue?: Buffer }[] | undefined,
-                tags: doc.tags.map((tag) => tag.id),
-                ...(doc.encryptedValue
-                  ? {
-                      encryptedValue: doc.encryptedValue,
-                      references: doc.value ? getAllSecretReferences(doc.value).nestedReferences : [],
-                      secretValueBlindIndex: doc.value
-                        ? await generateSecretBlindIndex(Buffer.from(doc.value))
-                        : undefined
-                    }
-                  : {
-                      encryptedValue: undefined,
-                      references: undefined
-                    })
-              }
-            }))
-          );
-
-          await fnSecretBulkUpdate({
-            folderId: destinationFolder.id,
-            orgId: actorOrgId,
-            resourceMetadataDAL,
-            folderCommitService,
-            secretVersionDAL,
-            secretDAL,
-            tx,
-            secretTagDAL,
-            secretVersionTagDAL,
-            actor: {
-              type: actor,
-              actorId
-            },
-            inputSecrets: inputSecretsForUpdate
-          });
-        }
-
-        const createdSecretsGroupedByKey = groupBy(createdSecrets, (s) => s.key);
-        destinationSecretIdByKey = Object.fromEntries(
-          decryptedSourceSecrets.map((s) => {
-            // for created secrets, use the newly created ID
-            if (createdSecretsGroupedByKey[s.key]?.[0]) {
-              return [s.key, createdSecretsGroupedByKey[s.key][0].id];
-            }
-            return [s.key, destinationSecretsGroupedByKey[s.key]?.[0]?.id];
-          })
-        );
-
-        isDestinationUpdated = true;
-      }
-
-      // Next step is to delete the secrets from the source folder:
-      const sourceSecretsGroupByKey = groupBy(sourceSecrets, (i) => i.key);
-      const locallyDeletedSecrets = decryptedSourceSecrets.map((el) => ({ ...el, operation: SecretOperations.Delete }));
-
-      const sourceFolderPolicy = await secretApprovalPolicyService.getSecretApprovalPolicy(
-        projectId,
-        sourceFolder.environment.slug,
-        sourceFolder.path
-      );
-
-      if (sourceFolderPolicy && actor === ActorType.USER) {
-        // if secret approval policy exists for source, we create the secret approval request
-        const localSecretsIds = decryptedSourceSecrets.map(({ id }) => id);
-        const latestSecretVersions = await secretVersionDAL.findLatestVersionMany(sourceFolder.id, localSecretsIds, tx);
-        const approvalRequestDoc = await secretApprovalRequestDAL.create(
-          {
-            folderId: sourceFolder.id,
-            slug: alphaNumericNanoId(),
-            policyId: sourceFolderPolicy.id,
-            status: "open",
-            hasMerged: false,
-            committerUserId: actorId
-          },
-          tx
-        );
-
-        const commits = locallyDeletedSecrets.map((doc) => {
-          const { operation } = doc;
-          const localSecret = sourceSecretsGroupByKey[doc.key]?.[0];
-
-          return {
-            op: operation,
-            requestId: approvalRequestDoc.id,
-            metadata: doc.metadata,
-            key: doc.key,
-            encryptedComment: doc.encryptedComment,
-            encryptedValue: doc.encryptedValue,
-            skipMultilineEncoding: doc.skipMultilineEncoding,
-            secretId: localSecret.id,
-            secretVersion: latestSecretVersions[localSecret.id].id
-          };
-        });
-
-        await secretApprovalRequestSecretDAL.insertV2Bridge(commits, tx);
-      } else {
-        // if no secret approval policy is present, we delete directly.
-        await secretDAL.delete(
-          {
-            $in: {
-              id: locallyDeletedSecrets.map(({ id }) => id)
-            },
-            folderId: sourceFolder.id
-          },
-          tx
-        );
-
-        isSourceUpdated = true;
-      }
-
-      // update references to the moved secrets whenever the destination was updated directly.
-      // this ensures references are updated regardless of whether the source has an approval policy.
-      // the secrets now exist at the destination, so references should point there.
-      if (isDestinationUpdated) {
-        for await (const secret of decryptedSourceSecrets) {
-          const destinationSecretId = destinationSecretIdByKey[secret.key];
-          if (!destinationSecretId) {
-            // eslint-disable-next-line no-continue
-            continue;
-          }
-
-          await fnUpdateMovedSecretReferences({
-            orgId: actorOrgId,
-            projectId,
-            sourceEnvironment,
-            sourceSecretPath,
-            sourceFolderId: sourceFolder.id,
-            destinationEnvironment,
-            destinationSecretPath,
-            destinationFolderId: destinationFolder.id,
-            secretKey: secret.key,
-            secretId: destinationSecretId,
-            secretDAL,
-            secretVersionDAL,
-            folderCommitService,
-            folderDAL,
-            secretQueueService,
-            encryptor: ({ plainText }) => secretManagerEncryptor({ plainText }),
-            decryptor: ({ cipherTextBlob }) => secretManagerDecryptor({ cipherTextBlob }),
-            generateSecretBlindIndex,
-            tx
-          });
-        }
-      }
-    });
-
-    if (isDestinationUpdated || isSourceUpdated) {
-      await secretDAL.invalidateSecretCacheByProjectId(projectId);
-    }
+    sourceFolder,
+    destinationFolder,
+    isSourceUpdated,
+    isDestinationUpdated
+  }: TDispatchSecretMoveSideEffectsDTO) => {
     if (isDestinationUpdated) {
       await snapshotService.performSnapshot(destinationFolder.id);
       await secretQueueService.syncSecrets({
         projectId,
-        orgId: actorOrgId,
+        orgId,
         secretPath: destinationFolder.path,
         environmentSlug: destinationFolder.environment.slug,
         environmentName: destinationFolder.environment.name,
@@ -3484,7 +2999,7 @@ export const secretV2BridgeServiceFactory = ({
       await snapshotService.performSnapshot(sourceFolder.id);
       await secretQueueService.syncSecrets({
         projectId,
-        orgId: actorOrgId,
+        orgId,
         secretPath: sourceFolder.path,
         environmentSlug: sourceFolder.environment.slug,
         environmentName: sourceFolder.environment.name,
@@ -3500,11 +3015,45 @@ export const secretV2BridgeServiceFactory = ({
         ]
       });
     }
+  };
+
+  const moveSecrets = async (dto: TMoveSecretsDTO) => {
+    const result = await secretDAL.transaction((tx) =>
+      fnSecretMoveInTransaction({
+        ...dto,
+        tx,
+        permissionService,
+        kmsService,
+        folderDAL,
+        secretDAL,
+        secretVersionDAL,
+        secretTagDAL,
+        secretVersionTagDAL,
+        resourceMetadataDAL,
+        folderCommitService,
+        secretApprovalPolicyService,
+        secretApprovalRequestDAL,
+        secretApprovalRequestSecretDAL,
+        secretQueueService
+      })
+    );
+
+    if (result.isDestinationUpdated || result.isSourceUpdated) {
+      await secretDAL.invalidateSecretCacheByProjectId(dto.projectId);
+    }
+
+    await dispatchSecretMoveSideEffects({
+      projectId: dto.projectId,
+      orgId: dto.actorOrgId,
+      actor: dto.actor,
+      actorId: dto.actorId,
+      ...result
+    });
 
     return {
-      projectId,
-      isSourceUpdated,
-      isDestinationUpdated
+      projectId: dto.projectId,
+      isSourceUpdated: result.isSourceUpdated,
+      isDestinationUpdated: result.isDestinationUpdated
     };
   };
 
@@ -4158,6 +3707,7 @@ export const secretV2BridgeServiceFactory = ({
     getSecretVersions,
     backfillSecretReferences,
     moveSecrets,
+    dispatchSecretMoveSideEffects,
     getSecretsCount,
     getSecretsCountMultiEnv,
     getSecretsMultiEnv,
